Calibration
Changed format from u_int16_t to float for MasterBias, MasterDark, MasterFlat and calibratedLight because of overflow and underflow

+++ src/calibration/include/calibration.h
@@ -4,12 +4,12 @@
 #include <stdint.h>
 #include <sys/types.h>
 
-void masterBias(u_int16_t *bias_all, u_int16_t *master_bias, long width, long height, int bias_count);
+void masterBias(float *bias_all, float *master_bias, long width, long height, int bias_count);
 
-void masterDark(u_int16_t *dark_all, u_int16_t *master_bias, u_int16_t *master_dark, long width, long height, int dark_count);
+void masterDark(float *dark_all, float *master_bias, float *master_dark, long width, long height, int dark_count);
 
-void masterFlat(u_int16_t *flat_all, u_int16_t *master_bias, u_int16_t *master_flat, long width, long height, int flat_count);
+void masterFlat(float *flat_all, float *master_bias, float *master_flat, long width, long height, int flat_count);
 
-void calibrateLights(u_int16_t *light_all, u_int16_t *master_bias, u_int16_t *master_dark, u_int16_t *master_flat, u_int16_t *calib_all, long width, long height, int light_count);
+void calibrateLights(float *light_all, float *master_bias, float *master_dark, float *master_flat, float *calib_all, long width, long height, int light_count);
 
 #endif
